[Core] [Source] [Metadata] Refactor aside
