+++ core/datacap-service/src/main/java/io/edurt/datacap/service/service/BaseService.java
@@ -2,6 +2,7 @@ package io.edurt.datacap.service.service;
 
 import io.edurt.datacap.common.enums.ServiceState;
 import io.edurt.datacap.common.response.CommonResponse;
+import io.edurt.datacap.common.utils.CodeUtils;
 import io.edurt.datacap.common.utils.NullAwareBeanUtils;
 import io.edurt.datacap.common.utils.ReflectionUtils;
 import io.edurt.datacap.service.SelfException;
@@ -12,6 +13,7 @@ import io.edurt.datacap.service.entity.PageEntity;
 import io.edurt.datacap.service.entity.UserEntity;
 import io.edurt.datacap.service.repository.BaseRepository;
 import io.edurt.datacap.service.security.UserDetailsService;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.data.domain.Pageable;
 
 public interface BaseService<T extends BaseEntity>
@@ -35,6 +37,9 @@ public interface BaseService<T extends BaseEntity>
         }
         if (ReflectionUtils.hasField(configure, "user")) {
             ReflectionUtils.setFieldValue(configure, "user", UserDetailsService.getUser());
+        }
+        if (StringUtils.isEmpty(configure.getCode())) {
+            configure.setCode(CodeUtils.generateCode(false));
         }
         return CommonResponse.success(repository.save(configure));
     }
