+++ core/datacap-ui/src/views/layouts/metadata/components/MetadataContent.vue
@@ -1,0 +1,11 @@
+<template>
+
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue'
+
+export default defineComponent({
+  name: 'MetadataContent',
+})
+</script>
